mcpmementoauth/requirements.txt
fastmcp>=1.0.0
mcp>=1.0.0
python-dotenv>=1.0.0
openai>=1.0.0
aiofiles>=24.0.0
msal>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
cryptography>=41.0.0
python-jose>=3.3.0
